data-collection/requirements.txt
flask==2.0.1
pymongo==4.0.1
paho-mqtt==1.6.1
requests==2.24.0
python-dotenv==0.19.0
flask-cors==3.0.10
werkzeug==2.0.0
gunicorn==20.1.0
tinytuya==1.2.0
PyP100==0.0.18
